hou/requirements.txt
amqp==2.5.2
backports.functools-lru-cache==1.6.1
billiard==3.6.1.0
configparser==4.0.2
contextlib2==0.6.0.post1
cycler==0.10.0
Django==1.11.8
django-cors-headers==2.1.0
django-redis==4.10.0
djangorestframework==3.9.4
drf-extensions==0.5.0
importlib-metadata==1.1.0
kiwisolver==1.1.0
kombu==4.6.6
matplotlib==2.2.4
mongoengine==0.18.2
more-itertools==5.0.0
mysql-connector-python==8.0.11
MySQL-python==1.2.5
mysqlclient==1.4.6
olefile==0.46
pathlib2==2.3.5
Pillow==4.1.1
protobuf==3.11.1
pymongo==3.9.0
PyMySQL==0.9.3
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
redis==2.10.6
scandir==1.10.0
six==1.13.0
vine==1.3.0
zipp==0.6.0